fixed parent_id type and depth handling.

+++ src/db.rs
@@ -27,7 +27,7 @@ pub fn init_db() -> Result<Connection> {
 }
 
 //insert link -> DB 
-pub fn insert_link(conn: &Connection, url: &str, depth: usize, parent_id: i32) -> Result<i64>{
+pub fn insert_link(conn: &Connection, url: &str, depth: usize, parent_id: i64) -> Result<i64>{
     conn.execute("INSERT INTO link (url, depth, parent_id) VALUES (?1, ?2, ?3)",
     params![url, depth as i64, parent_id],
     )?;
